requirements.txt
fastapi[standard]
pydantic
asyncpg
pydantic-settings
sqlmodel
alembic
bcrypt
pyjwt